+++ use_in_case/lib/src/modifiers/busy_state_modifier.dart
@@ -1,6 +1,6 @@
 import 'package:use_in_case/src/event.dart';
-import 'package:use_in_case/src/interactor/modifiers/chained_modifier.dart';
 import 'package:use_in_case/src/invocator.dart';
+import 'package:use_in_case/src/modifiers/chained_modifier.dart';
 
 ///
 /// Implementation of a invocation modifier.