+++ use_in_case/lib/use_in_case.dart
@@ -2,9 +2,9 @@ library;
 
 export 'src/event.dart';
 export 'src/interactor.dart';
-export 'src/interactor/modifiers/busy_state_modifier.dart';
-export 'src/interactor/modifiers/chained_modifier.dart';
-export 'src/interactor/modifiers/modifier.dart';
-export 'src/interactor/modifiers/on_event_collector_modifier.dart';
-export 'src/interactor/modifiers/timeout_modifier.dart';
 export 'src/invocator.dart';
+export 'src/modifiers/busy_state_modifier.dart';
+export 'src/modifiers/chained_modifier.dart';
+export 'src/modifiers/modifier.dart';
+export 'src/modifiers/on_event_modifier.dart';
+export 'src/modifiers/timeout_modifier.dart';